@@ -1,1 +1,3 @@
-"""Tests for Gemini CLI."""
+"""Test fixtures package for comprehensive testing.
+Provides reusable mocks and fixtures for all test suites.
+"""
